lib: re-org and bug fixes
