+++ lib/utils.js
@@ -2,6 +2,7 @@
 
 const Sequelize = require('sequelize');
 const pluralize = Sequelize.Utils.pluralize;
+const { errors } = require('./constants');
 
 exports.formatOutput = (results, model, options) => {
 	let response = {};
@@ -41,4 +42,34 @@ exports.isSequelizeModel = (model) => {
 	}
 
 	return false;
+};
+
+exports.handleRequestError = (req, res) => {
+	return (e) => {
+		switch (e.name) {
+		case 'SequelizeForeignKeyConstraintError':
+			return res.status(400).json({ errors: [{ message: 'foreign key constraint error' }] });
+		case 'SequelizeValidationError':
+		case 'SequelizeUniqueConstraintError': {
+			const results = [];
+			e.errors.forEach((x) => {
+				results.push({
+					message: x.message,
+					field: x.path
+				});
+			});
+
+			return res.status(422).json({ errors: results });
+		}
+		case 'SequelizeDatabaseError':
+			return res.status(422).json({ errors: [{ message: e.message }] });
+		}
+
+		switch (e) {
+		case errors.notFound:
+			return res.status(404).json({ errors: [{ message: e.message }] });
+		}
+
+		return res.status(500).json({ errors: [{ message: e.message }] });
+	};
 };
