+++ package.json
@@ -4,7 +4,7 @@
   "description": "Create REST handlers for Sequelize models",
   "main": "index.js",
   "scripts": {
-    "test": "nyc mocha --exit tests"
+    "test": "nyc --reporter=lcov --reporter=text-summary mocha --exit tests"
   },
   "repository": {
     "type": "git",
